Fixed time-parameter for pthread_cond_timedwait() function. Proofed for macOS/Darwin but not for other Posix Systems.
This fix should solve the following issue:
If you try to start listening to a pubnub channel when no internet-connection is available, CPU usage goes to 100% and even does not go down when a connection is established.

Digged into that problem on our macOS Systems and found the problem:
The function pthread_cond_timedwait() expects a time-interval since january 1st 1970 as "abstime"-parameter, but currently the time-interval since the last reboot is passed. Therefore pthread_cond_timedwait() exits immediately and the endless-loop continues.
The macOS man-page for pthread_cond_timedwait() is really clear and says that gettimeofday() (CLOCK_REALTIME) is used to obtain the time that is compared to "abstime"-parameter.
The man-pages for other POSIX-systems are not that clear and they only speak of "absolute time". It also seems to be the CLOCK_REALTIME (and not CLOCK_MONOTONIC as pubnub code is currently using). I have checked this with the function pthread_condattr_getclock() on different posix-systems (online c-compilers).