Fix receiving data longer thana http_buf at a time
If we had received data from Pubnub server that is longer than the
size of our HTTP ("scratchpad") buffer, we would mishandle that. It
wouldn't show often, as default size of the buffer on Windows and
POSIX was 32000 octets, which is longer than most HTTP replies.

This was actually a regression introduced in commit
03dfd2cbf03cd998f8c4c361b5fdca7e01b83536.

This prompted a "small scale refactoring" of the buffer handling,
which is now a little simpler, more resilent, uses less memory and
fixes some other corner-cases which probably couldn't happen "as-is",
but, just like this regression, could have "awaken" with some other
"innocent looking" change.