Handle HTTP chunks larger than http_buff
This was merely being too eager to reject long chunks. If the chunk
is largert then `http_buff`, we can simply get it in several
iterations. Of course, it may turn out that the chunk _is_ too big to
fit in the reply buffer, but, no need to reject it beforehand.