A bunch of bug fixes and cleanups
* Several cleanups.

- Added some ASSERTs to check buffer access. These should not happen, but,
if there's some bug, they will, so: ASSERT()
- We now compile w/out warnings on MSVC - replaced usages of all
"insecure" C standard library functions
- a small issue in Windows makefile was fixed

* Bump version to 2.2.12

* Fixed and updated unit tests

* Fix receiving data longer than a http_buf at a time

If we had received data from Pubnub server that is longer than the
size of our HTTP ("scratchpad") buffer, we would mishandle that. It
wouldn't show often, as default size of the buffer on Windows and
POSIX was 32000 octets, which is longer than most HTTP replies.

This was actually a regression introduced in commit
03dfd2cbf03cd998f8c4c361b5fdca7e01b83536.

This prompted a "small scale refactoring" of the buffer handling,
which is now a little simpler, more resilent, uses less memory and
fixes some other corner-cases which probably couldn't happen "as-is",
but, just like this regression, could have "awaken" with some other
"innocent looking" change.

* Handle HTTP chunks larger than http_buff

This was merely being too eager to reject long chunks. If the chunk
is largert then `http_buff`, we can simply get it in several
iterations. Of course, it may turn out that the chunk _is_ too big to
fit in the reply buffer, but, no need to reject it beforehand.

* Fix compiler and static check warnings and GCC on Windows

Various fixes for compiler warnings and static check errors/warnings
reported by clang-analyizer (`scan-build`) and CppCheck.

Also, the Makefile for GCC on Windows was updated and now builds
basic examples.

* Bump version to 2.2.13

* Support OpenSSL 1.1.0+, `BIO_(s|g)et_conn_ip` ->
  `BIO_(s|g)et_conn_address`
* Some NMake preprocessor magic to detect the
  `.lib`s to use, as this has changed in 1.1.0+,
  at least in OpenSSL "distribution" we use

* Some fixes in Base64 module

* The encoded length was not correctly calculated in all cases
* Introduced a new helper function to calclulate the needed size of
  the char array for the encoded string
* Some warning fixed

* Small bugfixes and cleanups

A small bug in Base64 decoder was fixed.

Other than that, ASSERT noreturn warning fixed, some undefined
behavior removed and some style improvements.