Fix bug in timer lists and some cleanups
* Fix bug in timer lists

If the first context in the list (the next one to expire) was removed
from the list, the next one wouldn't get its "time to expire", but
would be left with "what it had", causing it to expire sooner.

* A bunch of cleanups

The focal cleanup here is the "final solution" for the
`PUBNUB_STATIC_ASSERT`, which is now trully portable and doesn't
introduce warning when the assert "holds" on all compilers and
checkers we tried (GCC, Clang, MSVC, CppCheck). Should probably work
that way on any "regular" C compiler, also.

Other than that, fixed (almost) all warnings reported by MSVC, its
`/analyze` and CppCheck.  Actually, a few CppCheck reports remain,
but those are either simply wrong (bad code analysis) or false alarms
(incomplete code analysis).