Fix HTTP CONNECT proxy w/HTTP Keep-Alive
We only need to look at the "haved URI" in the first HTTP GET after
HTTP CONNECT. Subsequent HTTP GETs, while the connection to
Pubnub server is "kept-alive", should ignore the saved path
(as it's the path from the _first_ HTTP GET).

Also, we now send the exact version of the C-core in `User-Agent`
HTTP header, taken "directly from the horse's mouth", that is,
from `pubnub_version_internal.h`.