Sanitize locks
Added a new lock for the timer list, instead of re-using the "watcher"
lock, which is for the array of sockets to watch. Before this, we
could get into a "reverse order of locking", which is a potential
deadlock. It would not manifest, because we do everything on the
same thread, but, it's better to make this "future safe".

Also, more granular locking gives to more reponsive system. Again,
in our case, since all is done on the same thread, it doesn't help,
but, again, it will help if we have a thread shism in the future.