Use MSG_NOSIGNAL on Linux
To avoid SIGPIPE on send(). There's still the potential SIGPIPE
on close() and this isn't supported in most other OSes, though it is
defined in POSIX.2008, so we might try to detect this support
in a more advanced way in the future.

In other news, fixed the remaining benign ASSERTs of calling
pbpal_send_status() while "not in sending state".

Also added some more logging for "interesting" events/code paths.