Fix some bugs in starting transaction and HTTP keep-alive
We used to set the state to `IDLE` before callback and only after the
callback set it `KEEP_ALIVE_IDLE`, if needed, but a transaction can
be started in the callback, which will change the state. This was
much more benign than it could have been, but, needs to be fixed
as it could become much worse after any significant change.

Also, if we start a new transaction in a kept-alive connection and it
fails (because, most likely, the connection was actually lost in the
meantime), we would leak the socket (sure, it's probably unusable
anyway, but, it would not have been closed).

Other than that, a warning on Windows was fixed (MSVS considers
strncpy() unsafe).