Fix OpenSSL 1.1.x callback interface and C++
One _has_ to provide a "hostname" to OpenSSL BIO functions,
_can't_ use `BIO_address` functions to provide a binary address.
So, if you have the IP address, convert it to a string.

Various C++ wrapper fixes, mostly related to GZIP support and
OpenSSL.

Basic support for Travis and Appveyor builds added. Only the
builds, no tests are being run yet.