Free debug memory when no more contexts allocated
Previously, if the user would call pubnub_init() and then
pubnub_free(), she would perceive that there is a memory leak.
Of course, if the `PUBNUB_ASSERT_LEVEL` was high enough.

But, that's just because we didn't release the memory allocated for
the helper debug array that keeps track of all allocated contexts. We
simply "wait" for the next alloc and reuse said allocated array.. That
was to optimise for the corner case when user would "constantly" keep
allocating and deallocating one (or a few) contexts. But, since this
is "highest debug level", it makes sense to slow it down to gain a
nicer perception of memory usage (no perceived leaks).

This also solves the problem of perceived leak at the end of
execution, because if the user had freed all contexts, we would
still not release this memory and some leak checker would report
a leak. Now, if she had freed all contexts, no leak would be
reported by the leak checker at and of execution.