CI builds fully functional
Both on Travis and Appveyor, including functional tests, which now
degrade gracefully on Pull Requests builds.

Also, improved handling of lost connection while in HTTP Keep-Alive
(idle). The number of scenarios in which we detect that connection
was lost (while we were in "keep-alive idle") and start a new one is
improved.