Reset tls connect timer (#58)
* Fix invalid detection of PNR_TIMEOUT w/OpenSSL

The TLS connection timeout was being detected even after the connection
was established. We now don't observe connection timeout once the
connection is established.

* The `PNR_IO_ERROR` is no longer deemed a reason to reset the
timetoken. The thing is, most often this is networking error, which
is no reason to reset the timetoken - quite the opposite, if we
try again with the old timetoken, we might get the messages that
were published in the meantime.

* remove an unused local variable

* Fix some printf-like type mismatch warnings