Use monotonic clock for TLS connection timeout
For TLS connection timeout, we used C standard time(), which
provides a wall-clock time. This prone to errors because of
time synchronisation, administrative changes (daylight savings) and
such.

Now we use a monotonic clock, which is not supported by the
C standard library, so we have to use platform-specific API.

We also simplified the related code which used to check for
`PUBNUB_TIMERS_API` with some `if`s. The thing is, it doesn't make
much sense to use OpenSSL w/out Timers API, as the savings are
negligible. We now have a static assert to make sure
`PUBNUB_TIMERS_API` is defined for OpenSSL.