pubnub_await() now reports PNR_TIMEOUT
This is, of course, for the sync interface (only).

Previously, the transaction timeout was used in the sync interface
only to limit the read/receive and write/send duration. But, now,
when using `pubnub_await()`, it will keep track of time spent in
the transaction, and, if it runs out, will automagically cancel
the transaction reporting bak `PNR_TIMEOUT` to the user.

Some header docs are updated to better reflect this. So is the
`pubnub_sync_sample.c`, which is actually updated for several
other reasons of  being outdated.