Publish via POST and DNS codec fixes
One can now publish via POST, using pubnub_publish_ex() in C and
corresponding wrappers in C++ and Qt.

Several smaller issues were discovered in DNS codec.  This prompted a
limited refactoring of DNS module, moving the codec to its own module
and making a thorough unit test for it. It has 100% line coverage and
100% non-ASSERT branch coverage (that is, only ASSERTs that cannot
actually fail in code as-is, but are there for maintenance, are not
covered).