GZIPed publish (via POST)
If publish is done via (HTTP) POST (instead of the default GET),
it can now be GZIPed, which can cause significant bandwidth savings
on large and compression-friendly messages. If compression is too
low, it will not be done, as it increases the CPU utilization for
little or no benefit. This threshold is adjustable by the user.

Also, a bug was fixed wherein publishing via POST _encrypted_ would
use a "dead" pointer for the encrypted data and thus, most likely,
send garbage.