Publish via POST (maybe GZIPed) and DNS codec fixes (#62)
* Publish via POST

One can now publish via POST, using pubnub_publish_ex() in C and
corresponding wrappers in C++ and Qt.

* GZIPed publish (via POST)

If publish is done via (HTTP) POST (instead of the default GET),
it can now be GZIPed, which can cause significant bandwidth savings
on large and compression-friendly messages. If compression is too
low, it will not be done, as it increases the CPU utilization for
little or no benefit. This threshold is adjustable by the user.

* DNS codec fixes

Several smaller issues were discovered in DNS codec.  This prompted a
limited refactoring of DNS module, moving the codec to its own module
and making a thorough unit test for it. It has 100% line coverage and
100% non-ASSERT branch coverage (that is, only ASSERTs that cannot
actually fail in code as-is, but are there for maintenance, are not
covered).