URL encode channel name(s)
In most REST APIs, the channel name(s) are part of the URL path
and need to be URL encoded, in case they have `/` `:` or other
"URL elements". Up until now, they weren't and it wasn't an
issuse most of the time, because channels usually have some
"human-readable" name. But, it had to be done, it just wasn't
high in the priority list, so it took a long time to finish.