Signals support (#83)
Support for the Signals is added. Signals are like messages except
that new signal on the channel pushes out (removes) the previous
signal (there is no history). Signals also can't have metadata and
compressing them is not supported.

On the receiving side, if using subscribe V2, user can see if the
received message is a signal or a (regular) published message. This 
prompted completing the support for subscribe V2 in C++ & Qt.

Also, pubnub_stop() is added to gracefully shutdown in the callback
interface.

Other than that, some smaller code cleanups (removing warnings
from various code checkers/analyzers).