fix assertion fail after 24 days (32 bit variant)
24 days =	2 073 600 000 milliseconds
int32 max =	2 147 483 647
25 days =	2 160 000 000 milliseconds

This variant fixes the problem of the overrun after 24 days.
It does so by not giving back the "uninitialized value" (0) on accident.

On edge cases it makes measurements wrong by +-1ms.
Also there is a rollover after 24 days.  With 32 bit this cannot be circumvented.

Why 0 as uninitialized value?  Because this copes well with "static" initialization.

Why -1 and not 1 as the alternate value when accidentally hitting 0?
Because -1 is a more descriptive "I am on the edge" than just 1.