Introduce connection timer and fix high CPU load. (#89)
Connection timer is started on DNS resolution and stopped on
connection establishment. It was separated as it could and
should be much shorted than the transaction timer and thus
user will be able to "catch" various kind of networking issues
quicker (rather than waiting for the whole transaction timer
to expire).

Also, when no context is running a transaction, the callback
interface would run a high (~100%) load on the "socket watcher"
thread. Essentially, it would busy-wait for some transaction
to actually start.

Some minor race conditions fixed also.