Fix handling of DNS retries and stale HTTP kept alive (#103)
* Fix handling of DNS retries and stale HTTP kept alive

Bottom line is that `pbnc_stop()` was not designed to handle
anything itself, only to "setup the stage" for the FSM to
handle things.  Unfortunately, in these two scenarios, some
handling had "slipped in".

* Fix unit test build

* Don't 'raise' PNR_CANCELLED in PBS_IDLE

* Update unit test

* Update another unit test

* Fix the update

* Fix race condition

pubnub_free() is not holding a lock (mutex), so it needs to call a function that will acquire it, instead of the one that assumes that the lock is held.