Support system name servers in async DNS client
The name server is chosen with the following precedence:

1. The server specified with `pubnub_dns_set_primary_server_ipv4()`
  (Assumes `PUBNUB_SET_DNS_SERVERS` is set, which is the default)
2. The server specified with `pubnub_dns_set_secondary_server_ipv4()`
  (Assumes `PUBNUB_SET_DNS_SERVERS` is set, which is the default)
3. The server configured by the system (e.g. DHCP or manually)
4. The hardcoded default server (usually `8.8.8.8`)

If any step fails to find a server, it will try the next in the
precendence list. Step 4 always succeeds.

When `PUBNUB_SET_DNS_SERVERS` is _not_ set, the name server configured
on the system is always chosen first.

This patch also fixes a minor memory leak in the Windows-specific name
server lookup function when an error occurs.